requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.10.0
attr==0.3.2
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-herald==0.3.0
django-widget-tweaks==1.5.0
frozenlist==1.8.0
geoip2==5.1.0
idna==3.11
jsonpickle==4.1.1
maxminddb==2.8.2
multidict==6.7.0
pillow==12.1.0
propcache==0.4.1
pycparser==2.23
PyMySQL==1.1.2
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
yarl==1.22.0
